dom/imptests/webapps/WebStorage/tests/submissions/Ms2ger/test_storage_local_in_js.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>storage_local_in</h1>
    <div id="log"></div>
    <script>
        test(function() {
            localStorage.clear();
            assert_false("name" in localStorage);
            localStorage["name"] = "user1";
            assert_true("name" in localStorage);
        });
        test(function() {
            localStorage.clear();
            assert_false("name" in localStorage);
            localStorage.setItem("name", "user1");
            assert_true("name" in localStorage);
            assert_equals(localStorage.name, "user1");
            localStorage.removeItem("name");
            assert_false("name" in localStorage);
        });
    </script>
 </body>
</html>
